suggestions restructure completed
