+++ src/App.js
@@ -1,13 +1,14 @@
 import './App.scss';
-import Header from './components/Header/Header.js';
+// import Header from './components/Header/Header.js';
 import jsonData from './assets/data.json';
 import { commentsCompare, commentsCompareReverse, upvoteCompare, upvoteCompareReverse  } from './helpers/sortFunctions';
 import { useEffect, useState } from 'react';
-import Main from './components/Main/Main';
+// import Main from './components/Main/Main';
 
 import NewFeedback from './layouts/NewFeedback/NewFeedback.js';
 import EditFeedback from './layouts/EditFeedback/EditFeedback.js';
 import FeedbackDetail from './layouts/FeedbackDetail/FeedbackDetail';
+import Suggestions from './layouts/Suggestions/Suggestions';
 
 function App() {
 
@@ -111,7 +112,8 @@ function App() {
             /> */}
             {/* <NewFeedback /> */}
             {/* <EditFeedback /> */}
-            <FeedbackDetail selectedFeedbackDetail={selectedFeedbackDetail} />
+            {/* <FeedbackDetail selectedFeedbackDetail={selectedFeedbackDetail} /> */}
+            <Suggestions setSelectedFeedbackDetail={setSelectedFeedbackDetail} setFeedbackDetailVisible={setFeedbackDetailVisible} suggestionData={suggestionData} statusArray={statusArray} setCardsVisible={setCardsVisible} setRoadmapVisible={setRoadmapVisible} featureArray={featureArray} selectedCategory={selectedCategory} setSelectedCategory={setSelectedCategory} dashboardOpen={dashboardOpen} setDashboardOpen={setDashboardOpen} selectedIndex={selectedIndex} setSelectedIndex={setSelectedIndex} />
         </div>
     );
 }
