+++ src/components/Main/Main.js
@@ -1,5 +1,5 @@
 import styles from './main.module.scss'
-import SuggestionCards from "../SuggestionCards/SuggestionCards.js";
+import SuggestionCards from "../../layouts/Suggestions/_subcomponents/SuggestionCards.js";
 import Dashboard from '../Dashboard/Dashboard.js';
 import Roadmap from '../Roadmap/Roadmap.js';
