+++ src/layouts/FeedbackDetail/_subcomponents/feedbackdetailmain.module.scss
@@ -3,5 +3,4 @@
     padding-bottom: 56px;
     padding-left: 24px;
     padding-right: 24px;
-
 }
