+++ src/layouts/Suggestions/_subcomponents/Hamburger.js
@@ -6,6 +6,7 @@ export default function Hamburger({dashboardOpen, setDashboardOpen}) {
     function handleClick(event) {
         event.preventDefault();
         setDashboardOpen(!dashboardOpen);
+        console.log(dashboardOpen)
     }
 
     return (
