+++ src/layouts/Suggestions/_subcomponents/emptycard.module.scss
@@ -6,6 +6,8 @@
     background-color: #FFFFFF;
     padding: 76px 24px;
     border-radius: 10px;
+    margin-left: auto;
+    margin-right: auto;
     &_heading {
         color: #3A4374;
         margin-top: 40px;
