+++ src/layouts/Suggestions/_subcomponents/suggestioncards.module.scss
