+++ src/layouts/Suggestions/_subcomponents/suggestionsheader.module.scss
@@ -1,0 +1,53 @@
+.header {
+    display: flex;
+    height: 72px;
+    background-image: url('../../../assets/suggestions/mobile/background-header.png');
+    background-repeat: round;
+    justify-content: space-between;
+    align-items: center;
+    padding: 0 24px;
+    color: #FFFFFF;
+
+    &_textContainer {
+        height: fit-content;
+    }
+
+    p {
+        opacity: 0.75;
+    }
+
+    svg {
+        cursor: pointer;
+    }
+}
+
+.subheader {
+    display: flex;
+    height: 56px;
+    background-color: #373F68;
+    color: #F2F4FE;
+    align-items: center;
+    justify-content: space-between;
+    padding: 0 24px;
+    &_selectedOption {
+        display: flex;
+        flex-direction: row;
+        position: relative;
+    }
+    &_selector {
+        cursor: pointer;
+    }
+    &_button {
+        cursor: pointer;
+        height: 40px;
+        width: 134px;
+        background-color: #AD1FEA;
+        border-radius: 10px;
+        text-align: center;
+        //this will be modified once the button component is created... it will be an element
+        h4 {
+            font-size: 13px;
+            color: #F2F4FE;
+        }
+    }
+}
