+++ src/layouts/Suggestions/_subcomponents/suggestionsmain.module.scss
@@ -1,0 +1,6 @@
+.suggestionsMain {
+    margin-left: 24px;
+    margin-right: 24px;
+    margin-top: 24px;
+    margin-bottom: 56px;
+}
